Add 64-bit four-spot (::) variable support
Introduce 64-bit unsigned integer variables using the four-spot (::)
prefix, extending the existing spot (16-bit) and two-spot (32-bit)
types. Includes double-hybrid (;;) arrays for 64-bit array storage,
and new constant syntax (## for 32-bit, #### for 64-bit).

Runtime: IntVariable uses ulong internally, ExecutionContext indexer
returns ulong with per-type overflow checks, Lib gains Mingle32,
Select64, Rotate64, and 64-bit unary ops. Mingle retains classic
16-bit truncation behavior for backward compatibility.

New error messages:
- E533 "64 BITS SHOULD BE ENOUGH FOR ANYONE"
- E534 "YOU CAN'T MINGLE WITH SOMEONE WHO ISN'T YOUR TYPE"

126 tests passing, all existing samples compile and run correctly.

+++ intercal.runtime/sicklib.cs
@@ -671,7 +671,7 @@ public class syslib : System.Object
 
         label_1001:
         {
-            uint depth = frame.ExecutionContext[".5"];
+            uint depth = (uint)frame.ExecutionContext[".5"];
             if (depth > 0)
             {
                 frame.ExecutionContext.Resume(depth);
@@ -795,7 +795,7 @@ public class syslib : System.Object
 
         label_1023:
         {
-            uint depth = frame.ExecutionContext[".3"];
+            uint depth = (uint)frame.ExecutionContext[".3"];
             if (depth > 0)
             {
                 frame.ExecutionContext.Resume(depth);
@@ -1232,7 +1232,7 @@ public class syslib : System.Object
 
         label_1504:
         {
-            uint depth = frame.ExecutionContext[".6"];
+            uint depth = (uint)frame.ExecutionContext[".6"];
             if (depth > 0)
             {
                 frame.ExecutionContext.Resume(depth);
